Extract the analyzers from VSCode extension.
This is probably more maintainable and adds the benefit of new analyzers
like CFamily.

However this is a paradigm shift from the current implementation.
The current implementation downloads the analyzers when the
corresponding parameter is enabled.

In this new implementation, all servers are present and enabled out of the box.
This makes things simpler, and removes the needs for subpackages for every
analyzer.

But this also means heavy change in the way this package is configured.

Additionally, many sonarlint LSP extensions have been added/modified.
I based my work on the source code of the typescript front end
of vscode's extension.